refactor: reorganise dnd lib
