+++ packages/core/lib/dnd/collision/dynamic/get-midpoint-impact.ts
