+++ packages/core/lib/dnd/collision/dynamic/index.ts
