+++ packages/core/lib/dnd/collision/dynamic/track-movement-interval.ts
