+++ packages/core/lib/dnd/dnd-kit/safe.ts
