+++ packages/core/lib/dnd/use-rendered-callback.ts
